+++ module-system/sov-modules-macros/tests/dispatch/derive_event.rs
@@ -1,11 +1,15 @@
 mod modules;
 use modules::{first_test_module, second_test_module};
-use sov_modules_api::macros::DefaultRuntime;
-use sov_modules_api::{DispatchCall, Spec, Event, Genesis, MessageCodec};
+use sov_modules_api::{DispatchCall, Event, Genesis, MessageCodec, Spec};
 use sov_test_utils::TestSpec;
 
-#[derive(Genesis, DispatchCall, Event, MessageCodec, DefaultRuntime)]
-#[serialization(serde::Serialize, serde::Deserialize, borsh::BorshDeserialize, borsh::BorshSerialize)]
+#[derive(Default, Genesis, DispatchCall, Event, MessageCodec)]
+#[serialization(
+    serde::Serialize,
+    serde::Deserialize,
+    borsh::BorshDeserialize,
+    borsh::BorshSerialize
+)]
 struct Runtime<S: Spec> {
     pub first: first_test_module::FirstTestStruct<S>,
     pub second: second_test_module::SecondTestStruct<S>,