+++ module-system/sov-modules-macros/tests/rpc/expose_rpc_associated_types.rs
@@ -1,5 +1,5 @@
 use jsonrpsee::core::RpcResult;
-use sov_modules_api::macros::{expose_rpc, rpc_gen, DefaultRuntime};
+use sov_modules_api::macros::{expose_rpc, rpc_gen};
 use sov_modules_api::{
     Address, CallResponse, Context, DispatchCall, EncodeCall, Error, Genesis, MessageCodec, Module,
     ModuleId, ModuleInfo, Spec, StateValue, WorkingSet,
@@ -7,7 +7,7 @@ use sov_modules_api::{
 use sov_state::ZkStorage;
 use sov_test_utils::ZkTestSpec;
 
-pub trait TestSpec: 'static {
+pub trait TestSpec: Default + 'static {
     type Data: Data;
 }
 
@@ -98,12 +98,13 @@ pub mod my_module {
 use my_module::rpc::{QueryModuleRpcImpl, QueryModuleRpcServer};
 
 #[expose_rpc]
-#[derive(Genesis, DispatchCall, MessageCodec, DefaultRuntime)]
+#[derive(Default, Genesis, DispatchCall, MessageCodec)]
 #[serialization(borsh::BorshDeserialize, borsh::BorshSerialize)]
 struct Runtime<S: Spec, T: TestSpec> {
     pub first: my_module::QueryModule<S, T::Data>,
 }
 
+#[derive(Default)]
 struct ActualSpec;
 
 impl TestSpec for ActualSpec {